+++ internal/acceptance_tests/blocks/dictionary_explicit_on_compute.tf
@@ -1,0 +1,6 @@
+resource "fastly_service_dictionary" "test" {
+  service_id    = fastly_service_compute.test.id
+  version       = {{.SERVICE_VERSION}}
+  name          = "{{.DICTIONARY_NAME}}"
+  force_destroy = true
+}