+++ internal/provider/provider.go
@@ -82,6 +82,7 @@ import (
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/configstoreitems"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/customdashboard"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/customvcl"
+	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/dictionary"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/dnszone"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/domain"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/domainmanagement"
@@ -245,6 +246,7 @@ func (p *fastlyProvider) Resources(_ context.Context) []func() resource.Resource
 		configstore.NewResource,
 		configstoreitems.NewResource,
 		customdashboard.NewResource,
+		dictionary.NewResource,
 		dnszone.NewResource,
 		domain.NewResource,
 		domainmanagement.NewResource,
@@ -404,6 +406,7 @@ func (p *fastlyProvider) ListResources(_ context.Context) []func() list.ListReso
 		cdnacl.NewListResource,
 		cdnaclentries.NewListResource,
 		condition.NewListResource,
+		dictionary.NewListResource,
 		domain.NewListResource,
 		loggingbigquery.NewListResource,
 		loggingblobstorage.NewListResource,